pkg: Moving cpuallocator back to pkg/
The cpuallocator contains only generic functions so moving
it back under pkg/ directory.
