+++ cmd/topology-aware/policy/topology-aware-policy.go
@@ -24,8 +24,8 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 
 	"github.com/containers/nri-plugins/pkg/config"
+	"github.com/containers/nri-plugins/pkg/cpuallocator"
 	"github.com/containers/nri-plugins/pkg/resmgr/cache"
-	"github.com/containers/nri-plugins/pkg/resmgr/cpuallocator"
 	"github.com/containers/nri-plugins/pkg/resmgr/events"
 	"github.com/containers/nri-plugins/pkg/resmgr/introspect"
