+++ pkg/cpuallocator/testdata/sysfs.tar.bz2
